requirements.txt
flask
google-api-python-client==1.5.5
google-cloud-storage
flask-cors